+++ TP-11/src/controllers/products.controller.js
@@ -33,11 +33,36 @@ export const pagination = async (req, res) => {
 
 export const addProduct = async (req, res) => {
     const newProduct = req.body
+    const { title, description, code, price, status, stock, category} = newProduct
     try {
-        const product = await pManager.addProduct(newProduct)
+        if (!title || !description || !code || !price || !status || !stock || !category){
+            throw new Error("Faltan campos")
+        }
+        //Valido los tipo de datos requeridos
+        if ( 
+            typeof title        !== "string" || 
+            typeof description  !== "string" || 
+            typeof code         !== "string" || 
+            typeof category     !== "string"
+        ){
+            throw new Error("Verifique que title, description, code o category sean de tipo string")
+        }
+        if ( 
+            typeof price !== "number" || 
+            typeof stock !== "number"
+        ){
+            throw new Error("Verifique que price o stock sean de tipo number")
+        }
+        
+        const product = await pManager.getProductByCode(code)
+        if(product){
+            throw new Error('Code ya existe')
+        } 
+            
+        const prodAdded = await pManager.addProduct(newProduct)
         const products = await pManager.getProducts() //productos actualizados
         io.emit('updateproducts', products) //los envio por websockets al front
-        res.status(201).send({product})
+        res.status(201).send({status: "success", product: prodAdded })
     } catch(error){
         res.status(400).send({error: error.message})
     }
@@ -57,6 +82,11 @@ export const updateProduct = async (req, res) => {
     const { pid } = req.params
     const bodyToUpdate = req.body
     try {
+        const productExist = await pManager.getProductById(pid)
+        if(!productExist){
+            throw new Error("No existe el producto que quiere actualizar")
+        }
+
         const statusUpdate = await pManager.updateProduct(pid, bodyToUpdate)
         const products = await pManager.getProducts() //productos actualizados
         io.emit('updateproducts', products) //los envio por websockets al front
@@ -67,8 +97,13 @@ export const updateProduct = async (req, res) => {
 }
 
 export const deleteProduct = async (req, res) => {
-    const pid = req.params.pid
+    const { pid } = req.params
     try {
+        const productExist = await pManager.getProductById(pid)
+        if(!productExist){
+            throw new Error("No existe el producto que quiere eliminar")
+        }
+
         const statusDelete = await pManager.deleteProduct(pid)
         const products = await pManager.getProducts() //productos actualizados
         io.emit('updateproducts', products) //los envio por websockets al front
